add perfis ADMIN e USER no carregamento da aplicação   - add UsuarioService

+++ src/main/java/com/example/securityjwt/repository/PerfilRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PerfilRepository extends JpaRepository<Perfil, Integer> {
+public interface PerfilRepository extends JpaRepository<Perfil, Long> {
 }
